multi-video support #14 & more improvements
Fixed all m3u8 downloader warnings

Now when the video downloader fails to find the file size it
retries twice. If failed twice it starts to remove certain qualities and
retries with what's available.
- After some testing I have found that file size fails means the video just isn't available
even if it's been detected.
- It'll loop like this a couple times until the retries hit max.
- Might scrape for the qualities again.
- If this keeps failing then check wco for that video and see if you get errors.

Reduced max retries to 5.

The default download folder is now in /user_folder/zen_downloads/
- Also fixed an issue where it wasn't attempting to create these folders.

Added support for episodes with 2 videos.
- Works in the same thread, so you must wait for the 2nd one to complete if found.
- Has the same name as the first, but it's append with "-01" such as Shimo Episode 1-01 (720p).mp4
- Works the same as the first video download on it's seperate coroutine including it's own retries etc.
# it's name is appended like that for sorting