1.1.13
Now if the random series aren't enabled, the console fills up the rest of the height.

Gave the console a border and fixed up some internal ui issues.

Added a pop up thats follows the mouse for random series (Experimental).
- shows the images much bigger and shows the title.
- experimental due to window resizing making it flicker sometimes.

Fixed the padding on the main window and removed the passing of the padding to the content scope.

Added support for .rpm update parsing.

UpdateWindow now parses github markdown.