1.1.15 (v2)
The Downloader's url bar now has a clear button and requests focus on launch.

Now instead of just opening the DownloadConfirm window, everything checks the database
for the cached series and attempts to download it first. It will then open the window or throw an error.

Adjusted the new DownloadThread to handle stopping/completions better.

Now regardless of what sorting method is chosen in downloads, it will show
the ones that aren't completed at the top.

LAST_DOWNLOAD is now updated for anything you have typed inside the
downloaders url bar.

Fixed an issue with url updater when the days are in the negatives.

Added new url: https://www.wcoflix.tv

Added Scroll To Start & Scroll To End options to random series right click.