Testing the uber jar for all platforms
I have made significant progress towards creating a jar file that will on normally through java instead of creating different installers.

I have built it on windows and it successfully runs on fedora so far which is major progress.

The only issue right now is with certificates due to the jars using the platforms jdk. I don't want that to be an issue.