1.2.4 | JAR Restructure
God I hope this works.

The entire project is now a multi-module project.

Every single file has been changed, so the github update will be huge.

Created a new JAR launcher to handle updates for the new JAR packages.

Updated the gradle tasks for better JAR creation.

Updated the functions to open files and links for be more universal.

Fixed the Update Window to handle the new JAR packages.

Updated the README for the new JAR packages.